Refactor to place every subplot related method in subplot.py
